complete changes to split of server and client
changed names for clarification of use. changed the app names to demd
for the server and dem for the client and changed the name of the
servers source file to daemon.c.  change the terminology for the cli to
be verbs and objects.

also added the ability to use shorthand for the cli and completed the
remaining json verbs for editing the configuration.