finished building the interface list
the interface list is based on config file found in PATH_NVMF_DEM_DISC.
each interface contains a list of controllers based on matching the
transports using trtype, address family, address, and netmask read from
the json.config file. each controller has a list of subsystems including
the nqn and whether the subsystem is open to all hosts or not.
subsystems also contains a list of host with their nqn and the access
they have for that subsystem.

also corrected infinate loop in xport_loop function.  it would loop,
not moving to the next controller, if the nvme-fabrics module is not
loaded.