change threading model for host requests
pthreads leaks memory on each thread creation/deletion so instead
of creating a thread for each host request for its log pages,
create a thread per interface for handling all hosts request by
having the listener thread place the endpoint on a queue that the
processing thread will pull and process, in sequence. this reduces
the number of threads created and manages the pthread issue in a
known way.

also added a file to the files directory for commands use in gdb
specifically to show how to disable SIGTERM events from stopping
the debugger.