wait for host thread for finish before killing listener thread
occationally valgrind would catch use of memory but the host
thread on exist. this what caused by the listener thread creating
the host thread detached. changing this to do a join and noting
if the event that caused the thread to stop was a shutdown even for
the endpoint if connected allow this memory to be used properly and
not freed before the child thread is done.