Fix opcode values so get_domain_nqn works again
it turns out that opcode have a syntax. even opcode do gets where
odd opcode do sets.  when the get_domain_nqn was changed to 0x21 it
no longer wrote the data or the driver did not copy the data.

also these opcode should NOT be in nvme.h and be specific to fabrics