demd: replace ofi with generic transport interface
created the concept of a transport ops interface (xp_ops)
and implemented the rdma functions to connect endpoints (xp_ep),
listen for connection and accept them (xp_pep), create memory
regions and get remote keys for rma commands (xp_mr), etc.

demd connects to target and collect log pages and listens for hosts
and provides the subset of provisioned log pages.

Currently, only subsystems with Allow Any Host set true are provided
in the provisioned log pages. Need to update the use of the memory
schema to properly check for individually provisioned hosts.

also fixed issue where check_host does not parse the json schema properly
and set a timeout (10 sec) on host connections.  This needs to be reworked
to be able to scale.