fix host_thread to handle multiple connections
the original function only handled one connection at a time and
waited until the host disconnected before servicing the next host.

given host connections may be persistent, this function now maintains
a list of hosts and walks that list to poll for messages to process
as well as adding new host connections as they come in.

if no messages are received from a host for a period of time,
that host will be disconnected and removed from the list.