pre for changes to demd to do oob/inb configuration
intent is to send configuration changes as they occur so the
command 'apply' is not needed.  as a result the functionality
to restart the daemon is not needed.

in review of this code, the globals for curl and json contexts
need not be global so all call that passed in a void pointer to
the context is not needed.

also in prep for oob configuration, curl module was added to the
dem daemon and the inculde moved to src/incl