add in-band and out-of-band configuration
Base functionality working but not complete.  Out-of-band port
configuration working but need subsystem configuration.  In-band
message sent from dem to a modified target.  Need demt to handle
in-band as well as redfish.  Need to add directory for patches to
modify upstream target driver to support Keep Alive, AENs, and
new in-band configuration.  Need to modify cli/http for new 'valid'
field on transports and devices.