demd: correct configuration setting for portid and allowed hosts
the configuration of portids only deleted the links to subsystems
if the portid changed.  the target would fail if any of the other
values of the address info changes since there where subsystems
using the portid.  this change is to remove all links before
changing anything for the portid and then relink the subsystems
after the change has been make.

removing a host from a subsystems allowed hosts list now walks
the list of subsystems to see if it is still used and deletes the
host from the target if no subsystems are using that host