separate supervisory controller and discovery queues
supervisory controller interface supports both in-band and
out-of-band connecttions. in-band through either a specificed
interfaces or the first portid of the target.

multiple discovery queues are needed to support target that do not
provide all logpages on transport specific discovery queues and
where allow any host is not set on all subsystems.