dem-dc: fix missing Interfaces and NSDevices for inband managed targets
the issue was seen in the web interface that inband managed target did
not have either the Interfaces or the NSDevices sections as was seen for
out-of-band managed targets.  the issue was that for out-of-band, the json
data was being saved and the in memory structure being created.  for inband,
only the in memory structure was created.  new functions have been added to
take the create in memory structure and generate the json components so that
when the web interface did a get on a target, this info would be provided.