dem-mon/dem-hac: log page list cleanup
dem-mon was not removing log pages from the subsystem logpage list and
not free that memory when aen identified the logpage was deleted.

for both, add checking if the subsystem logpage list was empty after the
log pages was removed and removes the subsystem structure if no log pages
exist to ensure the messages are consistent.