dem-dc: create a single discovery queue for subsystems the allow any host
bug: no log pages fetched until both a portid and a subsystem are created

observation: a discovery queue was being created for every pair of portid
and subsystem but not for targets with management mode of local.

correction: create a single discovery queue when the portid is created to
be used for all subsystems the have allow any access and only create
specific discovery queues when a subsystem needs a specific hostnqn to for
access.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>