dem html: Enable client side routing
Client side routing will help mainting browser history to enable the
back/foreward browser buttons.

Added onpopstate callback function to parse the url and call loadDoc
function.  As a result the hash portion of the url is used to keep the
information also stored in the #objectType, #objectValue, and #Filter
elements in the #workspace on the page.

Also made back buttons on screen do back history but this added
complexity due to tooltips.  Added function onBack to close the tooltip
as well as doing the window.history.back() call.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>