dem suite: generate consistant discovery log change notifications
rework config functions that should respond to requests for
discovery log changes to send notifications and always send the
notifications even if the dem-em is not running or the target is
locally managed.

also correct requests for notifications in dem-ac and dem-dm to
only resent the request and not get the features again.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>