dem-em: add ops to allow single executable to support multiple target mgmt modes
since both in-kernel and spdk targets are supported, it makes more
sense to have a single executable support both/either target management
modes, i.e. configfs for in-kernel and rpc for spdk, then to have to
reconfigure the make files and rebuild for each, with the same file
names.

This adds a level of complexity in that the em compiled for both will
have an additional argument to specify which is to be used for a given
target.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>