Correct commit 6a9dbc535183122e72885a61b0258a0a07e4ded0
changes work with iwarp and ib for connect but iwarp still is
not recieving data, i.e. logpages.  issue with original commit
was increasing queue depth to 32.  since nvme includes a multiplier
for queue depth of 3, the max_send_wr result was not valid for
ib or iwarp.  returning the queue depth to 2 allows ib to receive
data, and iwarp to still connect.  iwarp issue will be addressed
in a later patch.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>